Trivial use of KR to run templated SPARQL queries.
This serves to illustrate the use of building queries from Mustache
templates, and our own KR for connecting to and executing the query.

For example:

   (query "sparql/test/connection.mustache" {:value 3}
          {:db-url "http://hostname:10035"
           :repository-name "repo"
           :username "user"
           :password "pass"})
   ;; => ({?/a 3})

Unfortunately, this also serves to illustrate how leaky KR is as an
abstraction library, since the client is forced to refer to, and bind,
parameters that are specific to a KB implementation.