Emit the column names as a header row.
The result set is returned as a sequence of maps, for which the key
ordering may change.  Thus the column ordering in the final resulset may
not match the order of the variables in the query's `select` clause.

Returning the header ensures that the consumer has an indication of the
ordering.