Improve query reusability.
As written the queries are specific to one type of identifier (in this
case: UniProt).  This violates the DRY principle as it forces us to
re-write the same queries for each of the sources that may name the
protein; forcing customers to use one specific identifier places an
unecessary burden on them, and is counter to the goals of the KB.

With this commit the queries have been rewritten to take a BIO
identifier rather than a source-specific identifier.  This makes them
generic, but requires the templating agent to first look up the unique
BIO ID denoted by the source-specific ID.