Translate source IDs into BIO IDs.
Complementing a rewrite of the query templates to be re-usable and not
specific to any one source ID, the source IDs that are provided as input
are now translated into BIO IDs before being substituted into the query
templates.

Where before this package was concerned only with shepherding query
parameters into the template, it now has a logical translation and
query-chaining functions.  This changes the way that the API is
constructed to enable this "interception" functionality without
introducing significant overhead for each new query.

The 1-to-1 mapping that used to exist between the query asked for by the
user and the template has been replaced by an API function, encapsulated
within which is the mapping/chaining logic needed to compile the query
that will ultimately be used to fulfil the user's request.

A naming convention allows these query handler functions to be
dynamically resolved and validated using the query name provided by the
user.  This convention is supported by a macro `define-interface-fn`
that generates both the API and supporting query function.  These
functions are now the API and this changes the way the queries that are
available to users are listed.  Consequently that function has been
temporarily dropped.