Add tests for queries and supporting functions.
Presently the tests really only validate that the syntax of the query is
correct.  The infrastructure to add test data to support the queries has
been created, but I'm in two minds as to whether its worth the effort to
create test data for all queries.  Testing against a particular test set
will not highlight any regressions due to a change in the way that the
store data are structured.  And there's little logic in the queries to
be exercised (as compared to what one might have in stored procedures).