Derive an external (protein) ID from a BIO id.
A guiding principle of the queries in this package is that a user should
be able to name the entity that they're looking for using the
nomenclature with which they're most familiar; for example: for proteins
IDs may be sources from UniProt, RefSeq, IRefWeb, etc.

To provide a consistent experience, the output returned by the queries
should also be guided by this principle.  Thus, rather than
returning (only) the KB's own BIO identifiers, we should map these to
identifiers in the same domain as the user's input.  This provides not
only a consistent user experience, but also ensures that they have the
ability to link together datasets produced by different queries.

This commit simply adds the mechanisms by which a BIO ID may be resolved
to a so-called "external ID".  This resolution is implemented using a
multimethod that may be extended to provide specific functionality for a
particular entity type.