Return ext IDs in the protein interaction query.
This first attempt at using the machinery for mapping a BIO ID into an
external ID (based on the IDs used by the user as input) is particularly
expensive as it requires an additional trip to the store *FOR EACH* BIO
ID in the result set.

This inefficiency will be address in a future commit.