Translate symbol when rendering CSV.
It's pretty much a given that any symbol in the resultset is going to be
the product of KR's IRI->symbol translation.  By rewriting the
separator, we can cheat in queries that take these same terms as input.
Queries can include a 'PREFIX' directive for the namespace (matching
what is already registered in the KB) and when these terms are copied
from the output of another query, everything will just magically work.